Fixed date output when data is missing dates
I thought the problem was initially with the parseDate
function that I wrote. But it turns out that it wasn't
the problem. I already added the tests to try and debug
it, so I figured I might as well keep them.

The problem is that structs in Go have zero values, and
since time.Time is a struct, that means that it has a
zero value (0001-01-01 00:00:00 +000 UTC). So when a date
value in the data was missing, the date would come out as
01/01/01.

So in order to fix the problem, I made the time.Time values
on the NFIPCommunity struct pointers. This allows for date
values to be missing from the data. If we haven't parsed
a date successfully, then we leave the value nil; otherwise,
we set it the address of the successfully parsed date.

In order to further accommodate this change, each date needed
to have its own variable. Without it, subsequent date parses
would just overwrite the last time.Time that was parsed, and
all of the date values on the NFIPCommunity struct would just
point to the same time.Time value.